Fix the build on linux.
GCC and CLang are strict with the scoped_ptr internal pointer.
The PassAs must used the right cast.

An other issue is related to using STL with scoped_ptr.
STL classes make some constructor copy in our back which are not allowed
with the constructor of scoped_ptr.

R=fdoray@chromium.org, fdoray

BUG=

Review URL: https://codereview.appspot.com/61560049