Add an error when decoding an unsupported CsComp event.
There is a compact/compressed version of CSwitch which seems undocumented.
Until we get more information on this payload, we should output an error
to avoid chasing the context switch issue again.

% xperf -providers k

 COMPACT_CSWITCH: Compact Context Switch
[...]
 Diag           : PROC_THREAD+LOADER+DISK_IO+HARD_FAULTS+DPC+INTERRUPT+CSWITCH+PERF_COUNTER+COMPACT_CSWITCH

An other idea is to send the raw payload.
I believe it would mislead the libtrace users.

R=fdoray@chromium.org, fdoray

BUG=

Review URL: https://codereview.appspot.com/65270043